+++ package.json
@@ -52,6 +52,7 @@
     "dotenv": "^16.4.5",
     "nodemon": "^3.1.0",
     "patch-package": "^8.0.0",
-    "typescript": "^5.4.5"
+    "typescript": "^5.4.5",
+    "vitest": "^3.0.6"
   }
 }
