+++ packages/adapters-library/src/adapters/integration.test.ts
@@ -7,8 +7,8 @@ import { bigintJsonParse } from '../core/utils/bigintJson'
 import { kebabCase } from '../core/utils/caseConversion'
 import { logger } from '../core/utils/logger'
 import { DefiProvider } from '../defiProvider'
-import { getInvalidAddresses } from '../scripts/addressValidation'
-import { protocolFilter } from '../scripts/commandFilters'
+import { getInvalidAddresses } from '../core/utils/addressValidation'
+import { protocolFilter } from '../core/utils/parameter-filters'
 import { RpcInterceptedResponses, startRpcMock } from '../tests/rpcInterceptor'
 import { IProtocolAdapter } from '../types/IProtocolAdapter'
 import { TestCase } from '../types/testCase'
